% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_w.R
\name{check_availability}
\alias{check_availability}
\title{This function checks if all relevant python modules are available}
\usage{
check_availability()
}
\description{
This function checks if all relevant python modules are available
}

