% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_w.R
\name{check_scorer}
\alias{check_scorer}
\title{secondary function for the 'FuzzExtract' class}
\usage{
check_scorer(scorer, DECODING)
}
\description{
secondary function for the 'FuzzExtract' class
}
\keyword{internal}
