% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_w.R
\name{is_python2}
\alias{is_python2}
\title{This function returns TRUE if python2 is installed and used in the OS}
\usage{
is_python2()
}
\description{
This function returns TRUE if python2 is installed and used in the OS
}
\keyword{internal}
